Added textfield background colors
This is mainly in the AddEditRecipeView, where the textfield will be red unless the user types in something. This is important as The Add/Edit recipe button is disabled unless the user completes set TextFields.
I wanted to make sure that the user will not be confused as to why the button does not work, now they are clearly red and awaiting some input.