Integrated CloudKit successfully.
CloudKit now works in real time, and with no issues. I have had some trouble with cloudkit as it was crashing my app and i didn't know why. I found out the trouble was with Constraints on an entity as CloudKit does not support unique constraints.
I had "title" constraints on Tag as I wanted titles to be unique. Now I will be trying a Workaround to see if I can make Tag.title unique not just in their own entity but in the whole app.