Added Button to delete tags in SearchView
The .onDelete(perform: deleteFunc) was changed to a Delete button that is bound to the multi-selection of tags.
The previous code was making the UI over-loaded with red buttons next to each tag as the list had an EditMode.
To avoid that and making the UI more simple I added the delete button, where you can now select multiple tags to delete as well.
Also added localisations for the new alert messages.